+++ app/src/main/java/com/tgyuu/baekyoung_i/main/navigation/BaekyoungNavHost.kt
@@ -4,13 +4,14 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
+import com.tgyuu.baekyoung_i.auth.navigation.authNavigationRoute
 import com.tgyuu.baekyoung_i.auth.navigation.authScreen
 
 @Composable
 fun BaekyoungNavHost(
     navController: NavHostController,
     modifier: Modifier = Modifier,
-    startDestination: String,
+    startDestination: String = authNavigationRoute,
 ) {
     NavHost(
         navController = navController,
